requirements.txt
alembic==1.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pyTelegramBotAPI==3.6.6
pytest==5.3.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
validators==0.14.1
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==2.0.1
